Fix naming

+++ src/core/copy.spec.ts
@@ -8,8 +8,8 @@ import { array } from './array';
 
 describe('(v) copy', () => {
   it('should create an immutable copy of class', () => {
-    const f32 = new Float64Array([1, 2, 3, 4]);
-    const original = array(f32);
+    const f64 = new Float64Array([1, 2, 3, 4]);
+    const original = array(f64);
     const cp = original.copy();
 
     notStrictEqual(original, cp);
@@ -18,8 +18,8 @@ describe('(v) copy', () => {
   });
 
   it('should work as the static equivalent', () => {
-    const f32 = new Float64Array([1, 2, 3, 4]);
-    const original = array(f32);
+    const f64 = new Float64Array([1, 2, 3, 4]);
+    const original = array(f64);
     const cp = copy(original);
 
     notStrictEqual(original, cp);
